refactor: features grouped in one dir
