+++ src/lib/db/index.ts
@@ -36,7 +36,7 @@ import { FavoriteProjectsStore } from './favorite-projects-store';
 import { AccountStore } from './account-store';
 import ProjectStatsStore from './project-stats-store';
 import { Db } from './db';
-import { ImportTogglesStore } from '../export-import-toggles/import-toggles-store';
+import { ImportTogglesStore } from '../features/export-import-toggles/import-toggles-store';
 
 export const createStores = (
     config: IUnleashConfig,
