+++ src/lib/features/export-import-toggles/export-import-permissions.e2e.test.ts
@@ -1,9 +1,9 @@
 import {
     IUnleashTest,
     setupAppWithAuth,
-} from '../../test/e2e/helpers/test-helper';
-import dbInit, { ITestDb } from '../../test/e2e/helpers/database-init';
-import getLogger from '../../test/fixtures/no-logger';
+} from '../../../test/e2e/helpers/test-helper';
+import dbInit, { ITestDb } from '../../../test/e2e/helpers/database-init';
+import getLogger from '../../../test/fixtures/no-logger';
 import {
     DEFAULT_PROJECT,
     IContextFieldStore,
@@ -13,11 +13,11 @@ import {
     IProjectStore,
     IUnleashStores,
     RoleName,
-} from '../types';
-import { ImportTogglesSchema, VariantsSchema } from '../openapi';
-import { IContextFieldDto } from '../types/stores/context-field-store';
-import { AccessService } from '../services';
-import { DEFAULT_ENV } from '../util';
+} from '../../types';
+import { ImportTogglesSchema, VariantsSchema } from '../../openapi';
+import { IContextFieldDto } from '../../types/stores/context-field-store';
+import { AccessService } from '../../services';
+import { DEFAULT_ENV } from '../../util';
 
 let app: IUnleashTest;
 let db: ITestDb;
