+++ src/lib/features/export-import-toggles/export-import-service.ts
@@ -1,37 +1,37 @@
-import { IUnleashConfig } from '../types/option';
+import { IUnleashConfig } from '../../types/option';
 import {
     FeatureToggleDTO,
     IFeatureStrategy,
     IFeatureStrategySegment,
     IVariant,
-} from '../types/model';
-import { Logger } from '../logger';
-import { IFeatureTagStore } from '../types/stores/feature-tag-store';
-import { ITagTypeStore } from '../types/stores/tag-type-store';
-import { IEventStore } from '../types/stores/event-store';
-import { IStrategy } from '../types/stores/strategy-store';
-import { IFeatureToggleStore } from '../types/stores/feature-toggle-store';
-import { IFeatureStrategiesStore } from '../types/stores/feature-strategies-store';
-import { IFeatureEnvironmentStore } from '../types/stores/feature-environment-store';
-import { IContextFieldStore, IUnleashStores } from '../types/stores';
-import { ISegmentStore } from '../types/stores/segment-store';
-import { ExportQuerySchema } from '../openapi/spec/export-query-schema';
+} from '../../types/model';
+import { Logger } from '../../logger';
+import { IFeatureTagStore } from '../../types/stores/feature-tag-store';
+import { ITagTypeStore } from '../../types/stores/tag-type-store';
+import { IEventStore } from '../../types/stores/event-store';
+import { IStrategy } from '../../types/stores/strategy-store';
+import { IFeatureToggleStore } from '../../types/stores/feature-toggle-store';
+import { IFeatureStrategiesStore } from '../../types/stores/feature-strategies-store';
+import { IFeatureEnvironmentStore } from '../../types/stores/feature-environment-store';
+import { IContextFieldStore, IUnleashStores } from '../../types/stores';
+import { ISegmentStore } from '../../types/stores/segment-store';
+import { ExportQuerySchema } from '../../openapi/spec/export-query-schema';
 import {
     FEATURES_EXPORTED,
     FEATURES_IMPORTED,
     IFlagResolver,
     IUnleashServices,
     WithRequired,
-} from '../types';
+} from '../../types';
 import {
     ExportResultSchema,
     FeatureStrategySchema,
     ImportTogglesValidateSchema,
-} from '../openapi';
-import { ImportTogglesSchema } from '../openapi/spec/import-toggles-schema';
-import User from '../types/user';
-import { BadDataError } from '../error';
-import { extractUsernameFromUser } from '../util';
+} from '../../openapi';
+import { ImportTogglesSchema } from '../../openapi/spec/import-toggles-schema';
+import User from '../../types/user';
+import { BadDataError } from '../../error';
+import { extractUsernameFromUser } from '../../util';
 import {
     AccessService,
     ContextService,
@@ -39,7 +39,7 @@ import {
     FeatureToggleService,
     StrategyService,
     TagTypeService,
-} from '../services';
+} from '../../services';
 import { isValidField } from './import-context-validation';
 import { IImportTogglesStore } from './import-toggles-store-type';
 import { ImportPermissionsService } from './import-permissions-service';
