+++ src/lib/services/index.ts
@@ -46,7 +46,7 @@ import { Knex } from 'knex';
 import {
     createExportImportTogglesService,
     createFakeExportImportTogglesService,
-} from '../export-import-toggles';
+} from '../features/export-import-toggles/createExportImportService';
 import { Db } from '../db/db';
 
 // TODO: will be moved to scheduler feature directory
