+++ src/lib/types/stores.ts
@@ -32,7 +32,7 @@ import { IFavoriteFeaturesStore } from './stores/favorite-features';
 import { IFavoriteProjectsStore } from './stores/favorite-projects';
 import { IAccountStore } from './stores/account-store';
 import { IProjectStatsStore } from './stores/project-stats-store-type';
-import { IImportTogglesStore } from '../export-import-toggles/import-toggles-store-type';
+import { IImportTogglesStore } from '../features/export-import-toggles/import-toggles-store-type';
 
 export interface IUnleashStores {
     accessStore: IAccessStore;
